refactor: centralize benefit list logic
